tests/requirements.txt
pyzmq
pyyaml
numpy
msgpack
tqdm==4.65.0
scipy==1.11.1
matplotlib==3.9.2
PyYAML==6.0.1
einops==0.6.1
compressai==1.2.4
pandas==2.0.2
requests==2.31.0
numpy==1.26.4
open3d
bitstream
torchvision
